added depth() to comptroller shim
