+++ contracts/shims/ComptrollerShim.sol
@@ -2,8 +2,27 @@
 pragma solidity ^0.8.7;
 
 import "../market/OverlayV1Comptroller.sol";
+import "../interfaces/IOverlayToken.sol";
+import "../interfaces/IUniswapV3Pool.sol";
+import "../libraries/UniswapV3OracleLibrary/UniswapV3OracleLibraryV2.sol";
+import "../libraries/FixedPoint.sol";
 
 contract ComptrollerShim is OverlayV1Comptroller {
+
+    using FixedPoint for uint256;
+
+    uint256 internal X96 = 0x1000000000000000000000000;
+
+    IOverlayToken public ovl;
+
+    address public ovlFeed;
+    address public marketFeed;
+    address public eth;
+
+    bool public ethIs0;
+
+    uint public macroWindow;
+    uint public microWindow;
 
     constructor (
         uint _impactWindow,
@@ -11,7 +30,13 @@ contract ComptrollerShim is OverlayV1Comptroller {
         uint _staticCap,
         uint _brrrrdExpected,
         uint _brrrrdWindowMacro,
-        uint _brrrrdWindowMicro
+        uint _brrrrdWindowMicro,
+        uint _priceWindowMacro,
+        uint _priceWindowMicro,
+        address _marketFeed,
+        address _ovlFeed,
+        address _ovl,
+        address _eth
     ) {
 
         impactWindow = _impactWindow;
@@ -20,12 +45,43 @@ contract ComptrollerShim is OverlayV1Comptroller {
         brrrrdExpected = _brrrrdExpected;
         brrrrdWindowMacro = _brrrrdWindowMacro;
         brrrrdWindowMicro = _brrrrdWindowMicro;
+        macroWindow = _priceWindowMacro;
+        microWindow = _priceWindowMicro;
+        marketFeed = _marketFeed;
+        ovlFeed = _ovlFeed;
+        ethIs0 = IUniswapV3Pool(_ovlFeed).token0() == _eth;
 
     }
 
-    function depth () internal view override returns (uint256) {
 
-        return staticCap;
+    function depth () internal virtual override view returns (uint256 depth_) {
+
+        uint32[] memory _secondsAgo = new uint32[](2);
+        _secondsAgo[0] = uint32(microWindow);
+        _secondsAgo[1] = 0;
+
+        ( int56[] memory _ticks, uint160[] memory _invLiqs ) = IUniswapV3Pool(marketFeed).observe(_secondsAgo);
+
+        uint256 _sqrtPrice = TickMath.getSqrtRatioAtTick(
+            int24((_ticks[1] - _ticks[0]) / int56(int32(int(microWindow))))
+        );
+
+        uint256 _liquidity = (uint160(microWindow) << 128) / ( _invLiqs[1] - _invLiqs[0] );
+
+        uint _ethAmount = ethIs0
+            ? ( uint256(_liquidity) << 96 ) / _sqrtPrice
+            : FullMath.mulDiv(uint256(_liquidity), _sqrtPrice, X96);
+
+        ( _ticks, ) = IUniswapV3Pool(ovlFeed).observe(_secondsAgo);
+
+        uint _price = OracleLibraryV2.getQuoteAtTick(
+            int24((_ticks[1] - _ticks[0]) / int56(int32(int(microWindow)))),
+            1e18,
+            address(ovl),
+            eth
+        );
+
+        depth_ = lmbda.mulUp(( _ethAmount * 1e18 ) / _price).divDown(2e18);
 
     }
