+++ tests/markets/comptroller/test_cap.py
@@ -10,19 +10,13 @@ def print_events(tx):
     for i in range(len(tx.events['log'])):
         print(tx.events['log'][i]['k'] + ": " + str(tx.events['log'][i]['v']))
 
-def test_cap_correct_eth_quote_in_ovl ():
-    pass
-
 def test_cap_to_half_because_print ():
     pass
 
-def test_cap_correct_brrrr_fade_positive ():
+def test_cap_correct_eth_quote_in_ovl ():
     pass
 
 def test_cap_above_static_cap_because_burn ():
-    pass
-
-def test_cap_correct_brrrr_fade_negative ():
     pass
 
 def test_cap_constraint_on_liquidity_vanish ():
